requirements.txt
asgiref==3.9.2
black==25.9.0
click==8.3.0
Django==5.2.7
django-debug-toolbar==6.0.0
django-environ==0.12.0
djangorestframework==3.16.1
mypy_extensions==1.1.0
packaging==25.0
pathspec==0.12.1
platformdirs==4.4.0
psycopg==3.2.10
psycopg-binary==3.2.10
pytokens==0.1.10
sqlparse==0.5.3
